+++ verify/config/config.exs
@@ -1,3 +1,5 @@
 use Mix.Config
 
-config :ua_inspector, database_path: Path.expand("../databases", __DIR__)
+config :ua_inspector,
+  database_path: Path.expand("../databases", __DIR__),
+  startup_sync: true