+++ config/config.exs
@@ -9,6 +9,5 @@ end
 if Mix.env() == :test do
   config :ua_inspector,
     database_path: Path.expand("../test/fixtures", __DIR__),
-    ets_cleanup_delay: 10,
     startup_sync: true
 end