+++ lib/ua_inspector/storage/server.ex
@@ -12,13 +12,12 @@ defmodule UAInspector.Storage.Server do
 
       @behaviour unquote(__MODULE__)
 
-      @ets_cleanup_delay_default 30_000
-      @ets_data_table_name __MODULE__.Data
-      @ets_lookup_table_name __MODULE__.Lookup
+      @ets_table_name __MODULE__
+      @ets_table_opts [:named_table, :protected, :set, read_concurrency: true]
 
       def init(_init_arg) do
         if Config.get(:startup_sync, false) do
-          :ok = reload_databases()
+          :ok = reload_database()
         else
           :ok = GenServer.cast(__MODULE__, :reload)
         end
@@ -26,45 +25,41 @@ defmodule UAInspector.Storage.Server do
         {:ok, nil}
       end
 
-      def handle_info({:drop_data_table, nil}, state), do: {:noreply, state}
-
-      def handle_info({:drop_data_table, ets_tid}, state) do
-        :ok = ETS.delete(ets_tid)
-
-        {:noreply, state}
-      end
-
       def handle_cast(:reload, state) do
-        :ok = reload_databases()
+        :ok = reload_database()
 
         {:noreply, state}
       end
 
       def list do
-        case ETS.fetch_data(@ets_lookup_table_name, @ets_data_table_name) do
-          nil -> []
-          ets_tid -> :ets.tab2list(ets_tid)
+        case :ets.info(@ets_table_name) do
+          :undefined ->
+            []
+
+          _ ->
+            case :ets.lookup(@ets_table_name, :data) do
+              [{:data, entries}] -> entries
+              _ -> []
+            end
         end
       end
 
-      defp reload_databases do
-        @ets_lookup_table_name = ETS.create_lookup(@ets_lookup_table_name)
-        old_ets_tid = ETS.fetch_data(@ets_lookup_table_name, @ets_data_table_name)
-        new_ets_tid = ETS.create_data(@ets_data_table_name)
+      defp create_ets_table do
+        case :ets.info(@ets_table_name) do
+          :undefined ->
+            _ = :ets.new(@ets_table_name, @ets_table_opts)
+            :ok
 
-        :ok = ETS.store_data_entries(read_database(), new_ets_tid)
-        :ok = schedule_data_cleanup(old_ets_tid)
-        :ok = ETS.update_data(@ets_lookup_table_name, @ets_data_table_name, new_ets_tid)
-
-        :ok
+          _ ->
+            :ok
+        end
       end
 
-      defp schedule_data_cleanup(ets_tid) do
-        Process.send_after(
-          self(),
-          {:drop_data_table, ets_tid},
-          Config.get(:ets_cleanup_delay, @ets_cleanup_delay_default)
-        )
+      defp reload_database do
+        datasets = read_database()
+
+        :ok = create_ets_table()
+        true = :ets.insert(@ets_table_name, {:data, datasets})
 
         :ok
       end