+++ lib/ua_inspector/downloader/readme.ex
@@ -1,21 +1,35 @@
 defmodule UAInspector.Downloader.README do
   @moduledoc false
 
+  require Logger
+
   alias UAInspector.Config
 
   @readme "ua_inspector.readme.md"
 
-  @doc """
-  Returns the path to the local copy of the README file.
-  """
+  @doc false
   @spec path_local() :: Path.t()
-  def path_local, do: Path.join(Config.database_path(), @readme)
+  def path_local do
+    _ =
+      Logger.info(
+        "UAInspector.Downloader.README.path_local/0 has been" <>
+          " declared internal and will eventually be removed."
+      )
 
-  @doc """
-  Returns the path of the README file distributed in priv_dir.
-  """
+    Path.join(Config.database_path(), @readme)
+  end
+
+  @doc false
   @spec path_priv() :: Path.t()
-  def path_priv, do: Application.app_dir(:ua_inspector, ["priv", @readme])
+  def path_priv do
+    _ =
+      Logger.info(
+        "UAInspector.Downloader.README.path_priv/0 has been" <>
+          " declared internal and will eventually be removed."
+      )
+
+    Application.app_dir(:ua_inspector, ["priv", @readme])
+  end
 
   @doc """
   Writes the informational README file if remote database is the default.
@@ -33,13 +47,15 @@ defmodule UAInspector.Downloader.README do
   end
 
   defp do_write do
-    dirname_local = Path.dirname(path_local())
+    path_local = Path.join(Config.database_path(), @readme)
+    path_priv = Application.app_dir(:ua_inspector, ["priv", @readme])
+    dirname_local = Path.dirname(path_local)
 
     unless File.dir?(dirname_local) do
       File.mkdir_p!(dirname_local)
     end
 
-    {:ok, _} = File.copy(path_priv(), path_local())
+    {:ok, _} = File.copy(path_priv, path_local)
     :ok
   end
 end