+++ CHANGELOG.md
@@ -17,6 +17,8 @@
     - Several functions are now declared internal and will result in a `Logger.info/1` message when called until they will be eventually removed:
         - `UAInspector.Downloader.prepare_database_path/0`
         - `UAInspector.Downloader.read_remote/0`
+        - `UAInspector.Downloader.README.path_local/0`
+        - `UAInspector.Downloader.README.path_priv/0`
 
 ## v1.0.0 (2019-04-20)